Rework WP_App Interface
- Remove duplicated code on all platforms.
- Rename function to improve code reuse in nanoBooter and nanoCLR.
