+++ src/CLR/WireProtocol/WireProtocol_App_Interface.c
@@ -6,28 +6,21 @@
 #include <nanoWeak.h>
 #include "WireProtocol_App_Interface.h"
 
-////////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////////
-// The functions below are the ones that need to be ported to client applications, 
-// tipically nanoBooter and CLR debugger
-//
-//
-/////////////////////////////////////////////////////////////////////////////////////////////////
+extern int Messaging_ProcessPayload(WP_Message* message);
 
-// provided as weak to be replaced by actual implementation in the client application
-__nfweak int WP_App_ProcessHeader(WP_Message* message)
+int WP_App_ProcessHeader(WP_Message* message)
 {
-    (void)(message);
+    // check for reception buffer overflow
+    if (message->m_header.m_size > sizeof(receptionBuffer))
+    {
+        return false;
+    }
 
-    // default to false
-    return false;
+    message->m_payload = receptionBuffer;
+    return true;
 }
 
-// provided as weak to be replaced by actual implementation in the client application
-__nfweak int WP_App_ProcessPayload(WP_Message* message)
+int WP_App_ProcessPayload(WP_Message* message)
 {
-    (void)(message);
-
-    // default to false
-    return false; 
+    return Messaging_ProcessPayload(message);
 }
