+++ targets/FreeRTOS/common/nanoCLR/CMakeLists.txt
@@ -7,7 +7,6 @@ set(TARGET_FREERTOS_NANOCLR_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR} CACHE INTER
 
 # append nanoCLR source files
 list(APPEND TARGET_FREERTOS_NANOCLR_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/CLR_Startup_Thread.c)
-list(APPEND TARGET_FREERTOS_NANOCLR_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/WireProtocol_App_Interface.c)
 
 # append target PAL source files
 list(APPEND TARGET_FREERTOS_NANOCLR_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/targetPAL_Events.cpp)
